switched underlying type of PieceMoveType enum from int to byte

+++ Move.cs
@@ -72,7 +72,7 @@ namespace LittleOwl {
         }
     }
 
-    internal enum PieceMoveType : int {
+    internal enum PieceMoveType : byte {
         Pawn = 0,
         Knight = 1,
         Bishop = 2,
